+++ src/app/latte-chat/letters/archive/letter/[id]/edit/page.tsx
@@ -1,7 +1,5 @@
-'use client'
+import LetterEditPage from '@/pages/latte-chat/letter/edit/ui'
 
-import LetterEditContainer from './LetterEditContainer'
-
-export default function LetterEditPage() {
-  return <LetterEditContainer />
+export default function LetterEditRoute() {
+  return <LetterEditPage />
 }
