--- src/features/letter-create/index.tsx
@@ -1,1 +1,0 @@
-export { default as LetterCreateFeature } from './containers/LetterCreateContainer'
