+++ src/pages/latte-chat/letter/create/ui.tsx
@@ -1,0 +1,5 @@
+import { LetterCreateFeature } from '@/features/letter-form'
+
+export default function LetterCreatePage() {
+  return <LetterCreateFeature />
+}
