+++ src/pages/latte-chat/letter/edit/ui.tsx
@@ -1,0 +1,5 @@
+import { LetterEditFeature } from '@/features/letter-form'
+
+export default function LetterEditPage() {
+  return <LetterEditFeature />
+}
